requirements.txt
xmltodict
beautifulsoup4
pandas
openpyxl
PyGithub
python-dotenv